backend/requirements.txt
Django
djangorestframework
django-cors-headers
chromadb
sentence-transformers
langdetect
PyPDF2
python-docx
requests
numpy
python-dotenv